Refactored H2_LDL factorization example file
